Guarantee that post construction signals are always sent to Services
In the past, if a module threw an exception while being constructed the Service system would see the pre construction signal but not a post construction signal. This lead to some difficulty with the message logger's internal state. This commit guarantees that if a pre construction signal is sent, a post construction signal will also be sent.