EventPrcocessor uses new SubProcess constructor and reducedProcessHistory from RunPrincipal
The SubProcess no longer takes a ProcessHistoryRegistry as an argument to its constructor since it now has its own internal copy. This avoids possible updates of a ProcessHistoryRegistry from different threads.
In addition, RunPrincipal now holds its reducedProcessHistoryID so the EventProcessor now uses it from the RunPrincipal and not the Source.