Added new Service signals to watch for a delayed get from the Event.
Timing based services want to know if a module has requested data from the Event and this request then spawns a delayed get either by running an unscheduled EDProducer or causing a delayed read from the Source. This change adds two new signals to the ActivityRegistry for this purpose: preModuleEventDelayedGetSignal_ and postModuleEventDelayedGetSignal_. At the moment we only need signals for Event requests since they are the only ones which can be delayed. Also, delayed file reads caused by objects like edm::Ref do not trigger this signal because the requisite ModuleCallingContext is not available from that call since the call directly accesses the edm::EventPrincipal and not the edm::Event.
Thanks go to Andrea Bocci for noticing the need for these signals.