Bug Fix for Exception Behavior
Fix bug which would cause problems after an unrelated
exception was thrown in a module constructor or the
service post module construction methods. In some cases,
the post method would have been called twice which could
in turn cause a seg fault in some services. If an exception
was thrown in the service post method, the process could
have been immediately terminated.