Migrate Sim Code to New Random Service Interface
Migrate most simulation code to use the new interface
of the random number generator service designed to work
with the multithreaded Framework. The main interface
change is to require a StreamID or LuminosityBlockIndex
argument to the getEngine function. In most cases,
this required moving code from the constructor to
the event or beginLuminosityBlock method. Then a pointer
to the engine was percolated or passed in some way
to the point where it was used.

Note OscarProducer was migrated earlier and there are a
few pieces of code I skipped over that will be migrated
in the near future. This pull request takes care of most
of the digitizers and the mixing modules.