Another unrelated unit test fix
Here is a fix for the standalone unit test failure.
Probably this does not help, but there was some
question whether CPPUNIT would run the two subtests
in the right order and this fix guarantees that.

I suspect the actual cause of the recent errors is
that the tests are being run on a machine with more cores
or that is faster and the cleanup is occurring earlier,
too early. Files that are still being used are getting
cleaned up earlier than before.

None of this has anything to do with the changes
in the original pull request