Introduce the concept of "graceful runtime".  After a fixed amount of walltime, the input source sends a stop signal on lumi transitions.
This allows us to set an expiration time, after which, the framework exits
successfully after the current lumi is finished.  We plan to experiment
with having the workflow management system keeping the current job
outputs and create a new job consisting solely of the unprocessed
lumis as a means of "chopping off the tail" and giving us more
deterministic runtime.