Do not throw an exception when ROOT reports a missing dictionary problem
In the case where ROOT prints a message about a missing collection
proxy we no longer throw an exception. This was needed to handle the
case of reading an old file using a release for which a class in that
file no longer exists. The framework doesn't need the exception since
it has other ways of detecting the missing dictionary and then take
appropriate action.