Add unit tests for edm::AssociationVector usage
Once the edm::AssociationVector dictionary problems were fixed, it was possible to use the class in the View test.
Added a pyroot reading test for edm::AssociationVector. Unfortunately, a bug in ROOT prevents reading from working correctly. This unit test is commented out for now but will be activated once the ROOT bug is fixed.