Changed framework timing report to be appropriate for threads
The tools for measuring CPU time actually measure CPU time for all cores being used by the process. Therefore recording such information per Event, Path, and Module makes no sense when multiple threads are running. We now only record CPU time for the event loop as a whole and report only wall clock time for all other measurements. We also now report both the wall clock time for the entire event loop but also the sum of the time for each all Streams. The former allows throughput to be measured while the later is useful for determining the average time it takes one Event to be processed.