Bug fix. Add function to get top level PSet
Add a new function that allows one to get a
top level ParameterSet by passing in a
ModuleDescription which is available in
all modules.

This is the Core part of the changes needed to
fix modules that were using the old
getProcessParameterSet function. This depends
on a single global static ParameterSetID.
It fails when used in SubProcesses would return
the ParameterSet for the wrong process/SubProcess.
The bug only affects jobs with SubProcesses.
This global was also problematic for multithreading
(had to be protected with a mutex).

More changes will follow this to convert
all uses of the old function to an alternative
and then eventually to delete it. Merging this
PR quickly will make it easier to test the others
quickly.