Bug fix. Fix moduleDescription function
Fix the moduleDescription function in stream
EDAnalyzers.  Apparently nothing ever called this
because no one reported the problem. If called
it causes a seg fault.  A necessary pointer
was not being initialized that points from
the stream module instance to the adaptor that
manages the stream instances. Also added something
to a unit test that exercises this function.

This bug fix is needed for the new
getProcessParameterSetContainingModule
function to work for stream analyzers,
which is needed to fix a different SubProcess
bug. I noticed this while testing the SubProcess
bug fix.