Make missing consumes calls an exception
This makes it an exception when a module has
not declared the products it gets with a consumes
function call. Previous to this it just resulted
in an error being printed out.

Fixed HcalDigisValidation which had recently added
a getByLabel without a consumes call. Also fixed
two Framework unit tests that also had problems with
consumes calls.

This has been tested to not cause problems in the
IB by scanning all the log files of a recent
IB for the error message. Framework unit tests
also pass.

While this should cause no IB test failures, there
probably will be failures of other code not tested
in the IBs which has not been migrated to declare
the products it consumes.