Make JobReport::inputFileOpened sufficiently thread-safe
Now that the mixing module runs multiple threads, we can get multiple threads each calling JobReport::inputFileOpened and trying to open files for the mixing. The code path which is called must be thread safe. Although the container used for this purpose was already thread safe, the way the item in the container was used was not safe.