Added Modifier.toReplaceWith
Added the method toReplaceWith to Modifier. This function is used to replicate the state of the fromObj into the existing toObj. This allows a larger conversion of the toObj than can be handled to the toModify method.
(cherry picked from commit def3dadf6e3046cd9b843e03f0c538ee41de8b19)