Removed direct access to product from ProductHolderBase
For thread safety, the status of the product must be checked before accessing the product since the status operates as the synchronizer. Providing direct access to the product from the base class makes it too easy to miss the checks on the status.
Now only the internals of the inheriting classes have access to the product directly. All other access must go through resolveProduct.