Moved locking shared resources out of DelayedReader and into InputProductResolver
By moving the lock to InputProductResolver we are able to check to see if while we were waiting for the lock another thread already retrieved the same data product so we do not have to call DelayedReader::getProduct again. This also makes the InputProductResolver::putProduct call safe from multiple threads since it is only called while in the lock.