Record if a Producer fails to put a data product
If a Producer fails to put a product we now change the ProductResolver to have a status of failed.
This has to be handled special for Run/LuminosityBlock since a data product might be put only in the end transition. To handle those cases we reset those particular ProductResolvers to be in the ready state.