Fix thread-safety issue in edmNew::DetSetVector
The call to dataSize() can happen while the vector m_data is
having its size change. Therefore it is not thread safe to call
m_data.size() under that circumstance. Therefore when running onDemand
we need to use an std::atomic to control access to the value.
When reading back from a source, onDemand is always false so
we do not have to worry about initializing m_dataSize when
reading back from storage.