Fix problem introduced by recent PR #17950 related to dropOnInput
The intent is that dropOnInput removes entries in the
ProductRegistry in addition to dropping the products.
The recent commit inadvertently disabled that for branches
that were already dropped.

And also this exposed the fact that GetterOfProducts requires
dictionaries for all products in the ProductRegistry, even the
dropped ones. And this should not be necessary and wastes a little
CPU, so that is fixed also.