Sort worker summary after all streams have been added
The worker summary was reporting incorrect values because the code assumed the ordering of modules in the list was the same from one StreamSchedule to the next. This was violated since StreamSchedules fill the container in the order from allModules() but then at the end would sort based on name. Therefore the first StreamSchedule called would reorder the list and the other StreamSchedules would assume it was the original order.
Now sorting is done only after all StreamScheduled have been called.