Require ESInputTag created with one string to contain a ':'
When constructing an ESInputTag using only one non empty string,
that string must now contain a ':' to denote the separation between
the module and data labels.
This is meant to catch a common error where the single string was
meant to only specify the data label, but was instead specifying
the module label.