Be consistent when replacing a module or Sequence in a Process
Previously, Sequences or Paths which contained a Sequence which contained the item to be replaced were sometimes expanded and sometimes only had the contained Sequence changed.
Now if the sub-Sequence is known to the Process, that sub-Sequence is modified but its parent Sequences or Paths are left unchanged. If the sub-Sequence is unknown to the Process, then the parent is expanded.