Fix friendly class name for nested types of class templates
The previous update in the logic missed the case that for
"Class<T>::Nested" the parsing in handleTemplateArguments() should
stop at the first colon and pass "Class<T>" to subFriendlyName
(because reTemplateArgs expects the name to end with ">").

Rather than messing with reTemplateArgs, the problem is fixed by
recognizing the colon after a class template as a break point in the
parsing, and remove the second colon as well.