Reset prefetchRequested and WaitingTaskList only if reset does not originate from early delete
Add tests for consuming (but not reading) a product after early
deletion. The tests show that resetting these members also for
early-delete-resets leads to another round of prefetching. In case the
producer is in a Task (i.e. UnscheduledProductResolver is used for the
product), an exception gets thrown even if the consumer does not read
the product.

Also update class name name in the early deletion exception message.