Fix infrequent missing LuminosityBlock
Under very heavy load where a thread is paused at just the right part of the function, the LuminosityBlock cache would be empty when the begin LuminosityBlock transition starts. The fix was to not have a local variable holding the std::shared_ptr<> to the LuminosityBlockPrincipal.
Additionally, removed the unnecessary LuminosityBlockProcessingStatus arguments from streamEndLumiAsync.