Avoid possibly running module after Event has finished
If a module is on multiple Paths, it is possible that if a thread pauses at just the wrong time, the paused thread might restart after an Event is finished processing and lead to the module running after the Event finishes. Trying to update the run status before adding to the WaitngTaskList avoids that possibility.