Fix EventSetup prefetching when using 1 thread
In order to avoid accidentally having an ED module run on the same thread as a synchronous EventSetup get is happening, all ES based tasks must run in their own tbb arena. In order for this to work when using 1 thread, we need to have a task in the default arena do a synchronous wait for all prefetched ES data products.