Legacy modules should not use ES shared resources
Having a legacy module acquire an ES shared resource and then do
a synchronous get asking for an ES module to use that same resource
leads to a deadlock. As ED modules are not allowed to share resources
with ES modules already, we just have to be sure legacy modules do
not take the ES resources either.