Implement edmtest::EventIDProducer and edmtest::EventIDValidator
EventIDProducer reads the EventID from the current event and copies it
into the Event as a data product.

EventIDValidator reads the EventID from the current event and compares
it to a data product read from the Event.

They can be used to validate that an object produced in a given event is
being read back in the same event.