Make edm::Hash compatible with std::ranges::less
Supporting std::ranges::less allows std::ranges::sort() to be used
with edm::Hash. The std::ranges::less requires all six comparison
operators to be defined. I decided to use the three-way comparison
operator as the base implementation because of why not. The equality
and non-equality operators had to be defined explicitly because the
three-way comparison operator is custom.